pull

+++ src/main/java/com/milkyway/dreamform/service/CommunityService.java
@@ -2,9 +2,11 @@ package com.milkyway.dreamform.service;
 
 import com.milkyway.dreamform.dto.CommunityDto;
 import com.milkyway.dreamform.model.Community;
+import com.milkyway.dreamform.model.LikeCheck;
 import com.milkyway.dreamform.model.UploadFile;
 import com.milkyway.dreamform.model.User;
 import com.milkyway.dreamform.repository.CommunityRepository;
+import com.milkyway.dreamform.repository.LikeCheckRepository;
 import com.milkyway.dreamform.repository.UserRepository;
 import lombok.AllArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -17,6 +19,7 @@ import java.io.IOException;
 import java.security.Principal;
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Optional;
 
 @AllArgsConstructor
 @Slf4j
@@ -25,6 +28,8 @@ public class CommunityService {
     private CommunityRepository communityRepository;
     private UserRepository userRepository;
     private ImageService imageService;
+    private LikeCheckRepository likeCheckRepository;
+
 
     public Page<Community> getCommunityList(Pageable pageable) {
         pageable = PageRequest.of(
@@ -68,6 +73,34 @@ public class CommunityService {
         return communityRepository.updateViewCounts(id);
     }
 
+    @Transactional
+    public void updateLikeCounts(Long id, String userName) {
+        Community community = communityRepository.findById(id).get();
+        User user = userRepository.findByUsername(userName).get();
+        Optional<LikeCheck> communityAndUser = likeCheckRepository.findByCommunityAndUser(community, user);
+        communityAndUser.ifPresentOrElse(
+                likeCheck -> {
+                    likeCheckRepository.delete(likeCheck);
+                    community.disCountLike(likeCheck);
+                    community.updateLikeCount();
+                }, () -> {
+                    LikeCheck like = new LikeCheck();
+                    like.mappingCommunity(community);
+                    like.mappingUser(user);
+                    community.updateLikeCount();
+                    likeCheckRepository.save(like);
+                }
+        );
+    }
+
+    @Transactional
+    public boolean likeCheck(Long id, String userName) {
+        Community community = communityRepository.findById(id).get();
+        User user = userRepository.findByUsername(userName).get();
+        Optional<LikeCheck> communityAndUser = likeCheckRepository.findByCommunityAndUser(community, user);
+        return communityAndUser.isPresent();
+    }
+
     public boolean idCheck(Principal principal, String userName) {
         return principal.getName().equals(userName);
     }
@@ -75,27 +108,34 @@ public class CommunityService {
     @Transactional
     public void deleteCommunity(Long id) {
         Community community = communityRepository.findById(id).get();
-        imageService.deleteFile(community.getUploadFile().getImageOriginal());
+        if(community.getUploadFile() != null && !community.getUploadFile().getImageOriginal().isEmpty()) {
+            imageService.deleteFile(community.getUploadFile().getImageOriginal());
+        }
         communityRepository.deleteById(id);
     }
+
+
 
     public List<CommunityDto> getBest() {
         List<Community> communityList = communityRepository.findAll();
         List<Community> bestCommunityList = new ArrayList<>();
         CommunityDto communityDto = new CommunityDto();
-        for(int i = 0; i < 5; i++) {
-            Integer max = -1;
-            int item = 0;
-            for(int j = 0; j < communityList.size(); j++) {
-                if(communityList.get(j).getViewCounts() > max) {
-                    max = communityList.get(j).getViewCounts();
-                    item = j;
+        List<CommunityDto> bestCommunityDtoList = new ArrayList<>();
+        if(!communityList.isEmpty() && communityList != null && communityList.size() >= 3) {
+            for(int i = 0; i < 3; i++) {
+                Integer max = -1;
+                int item = 0;
+                for(int j = 0; j < communityList.size(); j++) {
+                    if(communityList.get(j).getLikeCounts() > max) {
+                        max = communityList.get(j).getLikeCounts();
+                        item = j;
+                    }
                 }
+                bestCommunityList.add(communityList.get(item));
+                communityList.remove(item);
             }
-            bestCommunityList.add(communityList.get(item));
-            communityList.remove(item);
+            bestCommunityDtoList = communityDto.toBestDtoList(bestCommunityList);
         }
-        List<CommunityDto> bestCommunityDtoList = communityDto.toBestDtoList(bestCommunityList);
         return bestCommunityDtoList;
     }
 }
